Database.Esqueleto.Experimental.From.SqlSetOperation

Documentation

newtype SqlSetOperation a

Data type used to implement the SqlSetOperation language this type is implemented in the same way as a From

Semantically a SqlSetOperation is always a From but not vice versa

Since: 3.5.0.0

Constructors

SqlSetOperation

Fields

unSqlSetOperation :: NeedParens -> SqlQuery (a, IdentInfo -> (Builder, [PersistValue]))

class ToSqlSetOperation a r | a -> r where

Type class to support direct use of SqlQuery in a set operation tree

Since: 3.5.0.0

Methods

toSqlSetOperation :: a -> SqlSetOperation r

mkSetOperation :: (ToSqlSetOperation a a', ToSqlSetOperation b a') => Builder -> a -> b -> SqlSetOperation a'

Helper function for defining set operations @since 3.5.0.0

data Union a b

Deprecated: Since: 3.4.0.0 - Use the union_ function instead of the Union data constructor

Constructors

a `Union` b

Deprecated: Since: 3.4.0.0 - Use the union_ function instead of the Union data constructor

class Union_ a where

Overloaded union_ function to support use in both SqlSetOperation and withRecursive

Since: 3.5.0.0

Methods

union_ :: a

UNION SQL set operation. Can be used as an infix function between SqlQuery values.

class UnionAll_ a where

Overloaded unionAll_ function to support use in both SqlSetOperation and withRecursive

Since: 3.5.0.0

Methods

unionAll_ :: a

UNION ALL SQL set operation. Can be used as an infix function between SqlQuery values.

data UnionAll a b

Deprecated: Since: 3.4.0.0 - Use the unionAll_ function instead of the UnionAll data constructor

Constructors

a `UnionAll` b

Deprecated: Since: 3.4.0.0 - Use the unionAll_ function instead of the UnionAll data constructor

data Except a b

Deprecated: Since: 3.4.0.0 - Use the except_ function instead of the Except data constructor

Constructors

a `Except` b

Deprecated: Since: 3.4.0.0 - Use the except_ function instead of the Except data constructor

except_ :: (ToSqlSetOperation a a', ToSqlSetOperation b a') => a -> b -> SqlSetOperation a'

EXCEPT SQL set operation. Can be used as an infix function between SqlQuery values.

data Intersect a b

Deprecated: Since: 3.4.0.0 - Use the intersect_ function instead of the Intersect data constructor

Constructors

a `Intersect` b

Deprecated: Since: 3.4.0.0 - Use the intersect_ function instead of the Intersect data constructor

intersect_ :: (ToSqlSetOperation a a', ToSqlSetOperation b a') => a -> b -> SqlSetOperation a'

INTERSECT SQL set operation. Can be used as an infix function between SqlQuery values.

pattern SelectQuery :: p -> p

Deprecated: Since: 3.4.0.0 - It is no longer necessary to tag SqlQuery values with SelectQuery